requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-translate==3.9.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.6
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
